feat: Add Code Lint Checking and Black code style
* feat: Added lint checking

Adds lint checking for all code.  It is using the base from HACS with a few items removed.   I have also set this up to auto create a PR against what was submitted with any lint corrections when it fails

* Create pre-commit-config.yaml

* fix: Corrected Branch

* fix: Lint check on Pull Only

Currently this action will fail to create a PR to fix anything if you commit directly to master.  Only running it on PRs to start.

* Chore: latest pyaction version

* action fixes automatically created. (#131)

Co-authored-by: cdnninja <cdnninja@users.noreply.github.com>

* action fixes automatically created. (#135)

Co-authored-by: cdnninja <cdnninja@users.noreply.github.com>

* action fixes automatically created. (#136)

Co-authored-by: cdnninja <cdnninja@users.noreply.github.com>

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: cdnninja <cdnninja@users.noreply.github.com>