Hyundai USA: support a separate location fetch (Issue #196)
The location information provided in the cached vehicle status does not
include decimal precision.  In order to fix this, implement the
get_location method in the api for Hyundai USA.

Considering that there is a daily API limit, poll the get_location
method based on whether or not the odometer changed.  It will also keep
track of the last time that the location was retrieved form the server,
and only fetch a new location if the last update was more than an hour
ago.